mqtt-tfserving/requirements.txt
paho-mqtt
numpy
tensorflow
grpc